# developer-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Persona.ai Developer Platform API",
    "version": "1.0.0",
    "description": "REST + AG-UI API for integrating your product with Persona's agent infrastructure.\n\nEvery endpoint here is a server-to-server call — see the Integration Guide for where in your own stack to call each one from, and how the same endpoint serves both Project-level and per-end-user requests via the `x-persona-external-user-id` header."
  },
  "servers": [
    {
      "url": "https://api.persona.hasanraiyan.me",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "projectCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "<keyId>.<secret>",
        "description": "Developer Platform Project credential, minted via POST /api/v1/projects/{projectId}/credentials. Send as `Authorization: Bearer <keyId>.<secret>`. Optionally pair with the `x-persona-external-user-id` header to act on behalf of one of the Project's own external users. This credential is a server-side secret — see the Integration Guide."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/developer/mcps": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Create an MCP server (Project-owned or ExternalUser-owned)",
        "description": "Which kind gets created depends on the credential's asserted identity, same as Developer Agent creation.",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "transport",
                  "url"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "MCP server created"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "An MCP server with this exact name already exists"
          }
        }
      },
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Discover MCP servers",
        "description": "A genuinely separate code path from any Persona listing. Mcp has no isPublic/visibility field, so unlike Agent/Skill/Knowledge there is no separate \"public browse\" subset — a bare Project credential OR a ProjectRuntimeContext without scope=mine both see every MCP in this Project's own Domain, any owner type. scope=mine (ProjectRuntimeContext only) narrows to just that external user's own MCPs.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "mine"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of MCP servers"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Get an MCP server",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MCP server"
          },
          "404": {
            "description": "MCP server not found (or not visible to this Domain/Subject)"
          }
        }
      },
      "patch": {
        "tags": [
          "Developer"
        ],
        "summary": "Update an MCP server (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated MCP server"
          },
          "404": {
            "description": "MCP server not found or unauthorized"
          },
          "409": {
            "description": "Another MCP server with this name already exists"
          }
        }
      },
      "delete": {
        "tags": [
          "Developer"
        ],
        "summary": "Delete an MCP server (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MCP server deleted"
          },
          "404": {
            "description": "MCP server not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/usage": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "See what's using this MCP, before attempting to delete it",
        "description": "Answers \"what's using this\" proactively, rather than only finding out via a rejected delete call. `agents` is a preview (capped at 20); `agentCount` is the real total.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ agentCount, agents: [{ _id, name }] }"
          },
          "404": {
            "description": "MCP server not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/mcps/bulk-delete": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Delete multiple MCP servers in one call",
        "description": "Best-effort batch delete — one blocked/not-found id doesn't abort the rest. Up to 100 ids per request.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ deleted: string[], failed: [{ id, reason }] }"
          },
          "400": {
            "description": "ids missing, empty, or over 100 entries"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/test": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Test an MCP server's connection (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discovered tools/resources"
          },
          "404": {
            "description": "MCP server not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/resource": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Read an MCP resource (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource content"
          },
          "400": {
            "description": "Missing uri"
          },
          "404": {
            "description": "MCP server not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/call-tool": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Call an MCP tool (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "arguments": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool result"
          },
          "400": {
            "description": "Missing tool name"
          },
          "404": {
            "description": "MCP server not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/oauth/owner/authorize": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Start the owner OAuth flow (owner only)",
        "description": "No separate Developer callback route exists — the OAuth redirect URI is a single, pre-registered URL that completes via the existing Persona callback route regardless of who initiated the flow, since the signed state alone carries the initiating context.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorization URL"
          },
          "404": {
            "description": "MCP server not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/oauth/user/authorize": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Start the per-user OAuth flow (ProjectRuntimeContext only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-persona-external-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorization URL"
          },
          "400": {
            "description": "No asserted external user"
          },
          "404": {
            "description": "MCP server not found, wrong Domain, or not configured for per-user auth"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/oauth/user/status": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Get the calling Subject's own per-user connection status",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection status"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/oauth/user/connection": {
      "delete": {
        "tags": [
          "Developer"
        ],
        "summary": "Disconnect the calling Subject's own per-user connection",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disconnected"
          }
        }
      }
    },
    "/api/v1/developer/mcps/{mcpId}/oauth/owner/connection": {
      "delete": {
        "tags": [
          "Developer"
        ],
        "summary": "Disconnect the owner-mode connection (owner only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "mcpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner connection disconnected"
          },
          "404": {
            "description": "MCP server not found or unauthorized"
          }
        }
      }
    },
...
  },
...
}